requirements.txt
haystack-ai==2.5.1
anthropic-haystack==1.0.0
cohere-haystack==1.1.3
pinecone_haystack==2.0.0
streamlit==1.38.0
python-dotenv==1.0.1
